# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: NOZOMI 用车平台 API
  version: 0.1.0
  description: |
    接口定义先于实现：新增或修改接口时先改本文件，再写代码。
    `apps/api/src/openapi.test.ts` 会校验「已注册的路由」和本文件的 `paths` 完全一致。

    约定：
    - 出错时统一返回 `{ error: { code, message, details } }`（见 `ErrorResponse`）。
      唯一的例外是 `/health` 的 503：为了让监控直接看到是哪一项出了问题，仍返回 `HealthResponse`。
    - 所有接口共有的错误（不在每个接口下重复列出）：
      - 400 `BAD_REQUEST`：请求体不是合法 JSON、URL 里有非法的百分号编码、发来的根本不是 HTTP 请求
      - 400 `VALIDATION_FAILED`：参数校验未通过，`details.issues` 列出每一项
      - 404 `NOT_FOUND`：路径或请求方法不存在
      - 408 `REQUEST_TIMEOUT`：请求迟迟没有发完
      - 413 `PAYLOAD_TOO_LARGE`：请求体超过 1MB
      - 415 `UNSUPPORTED_MEDIA_TYPE`：请求内容类型不支持
      - 431 `REQUEST_HEADER_FIELDS_TOO_LARGE`：请求头超过 16KB
      - 500 `INTERNAL_ERROR`：未预期的服务端错误，`message` 不含内部细节

      其中 408、431 和「不是 HTTP 请求」的 400 在解析阶段就被拒绝，应答之后连接会被服务端关闭。
    - 金额用最小货币单位整数，百分比用基点整数。
    - 租户接口的 `tenant_id` 只从登录令牌读取；请求参数和请求体里的 `tenant_id` 一律被忽略。
    - `/platform/v1` 和 `/tenant/v1` 的字段名用 snake_case，时间用带时区的 ISO 8601。
      （`/health` 是 M0-05 定下的运维接口，字段名保持原样。）
    - 登录：平台员工用 `/platform/v1/auth/login`，租户用户用 `/tenant/v1/auth/login`，
      拿到访问令牌后放在请求头 `Authorization: Bearer <令牌>`。令牌 8 小时过期，没有刷新令牌，过期后重新登录。
      两类令牌互不通用：平台令牌访问租户接口、租户令牌访问平台接口，都是 401。
      停用账号、退出登录、凭重置令牌改了密码之后，已经发出的令牌立即失效。
      租户被暂停不影响它的用户登录（暂停的租户还要继续履约已有订单），`tenant.status` 会告诉前端当前状态。
    - 所有字符串输入都不能包含 NUL 字符（`\u0000`、查询串里的 `%00`），也不能包含孤立的代理字符（被截断的半个表情符号），
      否则 400 `VALIDATION_FAILED`。
...
      什么都没有改动，原样重试即可。主数据的写接口会先自动重试，极少会真的返回它。
...
servers:
...
paths:
...
  /tenant/v1/areas:
    get:
      tags: [区域]
      operationId: listAreas
      summary: 区域列表
      description: |
        需要操作 `area.read`（管理员、商品价格、只读）。只返回令牌所属供应商自己的区域，按最近修改从新到旧
        （修改时间相同的按编号，顺序稳定）。列表里的每一项不带图形的坐标（没有 `polygons`），只带块数。
        `total` 是符合筛选条件的总数，和这一页的内容读自同一时刻的数据。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          description: 关键字：任意一种语言的名称里包含它（不区分大小写；`%`、`_` 按字面匹配）
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: city_id
          in: query
          description: 只看这个城市的区域
          schema:
            type: string
            format: uuid
        - name: biz_type
          in: query
          schema:
            $ref: "#/components/schemas/AreaBizType"
        - name: status
          in: query
          description: 按状态筛选；不带或 `all` 表示全部
          schema:
            type: string
            enum: [active, disabled, all]
            default: all
      responses:
        "200":
          description: 一页区域
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AreaSummary"
                  next_cursor:
                    type: [string, "null"]
                  total:
                    type: integer
                    minimum: 0
    post:
      tags: [区域]
      operationId: createArea
      summary: 新增区域
      description: |
        需要操作 `area.manage`（管理员、商品价格）。必须带请求头 `Idempotency-Key`。
        - 所属城市必须是平台启用中的城市；创建后不能改。新增的区域直接是启用的。
        - 至少一块营运区；每一块是不带洞的简单多边形。多边形给 `geometry`；圆给 `circle`（圆心 + 半径），
          后端用 `@nozomi/domain` 的 `circleToRing` 算出 64 边的多边形保存——判断「点在不在里面」只认多边形。
        - 保存时图形被整理成：6 位小数、逆时针（第一个点不变）、首尾闭合。返回的就是整理后的。
        - 每块图形的 `seq`（「营运 1」里的 1）由后端分配。
        - 同一个供应商在同一个城市里不能有同名的区域（不分大小写、全角半角，不看是哪种语言）。
        - 上限：一个区域最多 50 块图形、一个多边形最多 1000 个点、合计最多 5000 个点；圆的半径 100 到 100000 米。
        写审计日志（`resource` = `area`，`action` = `create`）。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaCreate"
      responses:
        "201":
          description: 新增的区域（同一个幂等键再来时，原样返回第一次的这份应答）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
        "400":
          $ref: "#/components/responses/AreaInvalid"
        "409":
          description: |
            - `AREA_NAME_TAKEN`：这个城市已经有同名的区域
            - `MASTER_DATA_NOT_READY`：所属城市已被平台停用（`details.reason` = `CITY_DISABLED`）
            - `CONCURRENT_UPDATE`：和别人同时进行的修改撞上了，什么都没有改动，请重试
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "`IDEMPOTENCY_KEY_REUSED`：这个幂等键已经用于另一个内容不同的请求"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenant/v1/areas/{id}:
    get:
      tags: [区域]
      operationId: getArea
      summary: 查看一个区域
      description: 需要操作 `area.read`。带全部图形。别的供应商的区域是 404。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 区域
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
    put:
      tags: [区域]
      operationId: updateArea
      summary: 修改区域
      description: |
        需要操作 `area.manage`。必须用请求头 `If-Match` 带上当前的 `version`。
        名称、业务类型、图形都是**整体替换**：`polygons` 里没有的图形被删掉。
        - 已有的图形带上它的 `id`（编号不变；类别没变的话序号也不变）；新加的图形不带 `id`，序号按这一类保存前的最大序号往上数，删掉的序号不再用。
        - 把 `GET` 拿到的 `polygons` 原样提交回来是合法的（多出来的 `seq` 被忽略）。
        - 不能改城市：带了 `city_id` 且和现有的不同返回 409 `FIELD_LOCKED`。
        - 内容没有变化时原样返回，`version` 不变、不写日志。
        有变化时 `version` 加一并写审计日志（`action` = `update`，前后值只含变了的那几项）。区域停用时、所属城市被平台停用后也可以改。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaUpdate"
      responses:
        "200":
          description: 修改后的区域
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
        "400":
          $ref: "#/components/responses/AreaInvalid"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          description: |
            - `VERSION_CONFLICT`：`If-Match` 里的版本号不是最新的，`details.current_version` 是当前版本
            - `FIELD_LOCKED`：改了城市，`details.fields` = `["city_id"]`
            - `AREA_NAME_TAKEN`：这个城市已经有同名的区域
            - `CONCURRENT_UPDATE`：和别人同时进行的修改撞上了，请重试
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "428":
          $ref: "#/components/responses/VersionRequired"
    delete:
      tags: [区域]
      operationId: deleteArea
      summary: 删除区域
      description: |
        需要操作 `area.manage`。区域可以真的删除（连同它的图形），删了不能恢复；删除前的完整内容记在审计日志里（`action` = `delete`）。
        只是暂时不用的话请用停用。被已上架的商品引用时不能删除（409 `AREA_IN_USE`）——这条规则随商品（M1-03）一起生效，现在不会返回。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: 已删除
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          description: "`AREA_IN_USE`：有已上架的商品在用这个区域，`details.published_product_count` 是数量（M1-03 起）"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tenant/v1/areas/{id}/check-point:
    post:
      tags: [区域]
      operationId: checkAreaPoint
      summary: 自测：一个位置在区域的哪里
      description: |
        需要操作 `area.read`。按**已保存**的图形判断，禁行区优先；不管在不在里面都返回 200。正好压在边线上的位置算在里面。
        只看这一个区域自己的图形（禁行区只管它所在的区域）；停用的区域也能测。不写任何东西。
        判断用的是 `@nozomi/domain` 的 `locatePoint`，以后报价用的也是它。
      security:
        - tenantToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                  minimum: -90
                  maximum: 90
                lng:
                  type: number
                  minimum: -180
                  maximum: 180
      responses:
        "200":
          description: 判断结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AreaPointCheck"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /tenant/v1/map/config:
    get:
      tags: [区域]
      operationId: getTenantMapConfig
      summary: 地图底图的配置
      description: |
        任何已登录的供应商账号都能调（账号必须先修改密码时同样被拦）。返回这个环境的底图（栅格瓦片）配置；没有配置时 `tiles` 是 null，
        页面按「没有配置底图」显示，功能不受影响。
        值来自部署环境的配置（`MAP_TILE_URL_TEMPLATE` 等），内容和供应商无关。全站的内容安全策略 `img-src` 放行的就是这里的瓦片来源（同一个环境变量算出来的）。
        全站 `Referrer-Policy` 是 `no-referrer`；瓦片图片请按 `referrer_policy` 单独设置（多数底图服务要求带上来源域名）。
      security:
        - tenantToken: []
      responses:
        "200":
          description: 底图配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapConfig"
  /tenant/v1/dashboard/summary:
    get:
      tags: [区域]
      operationId: getTenantDashboardSummary
      summary: 供应商后台首页的数量
      description: |
        任何已登录的供应商账号都能调（账号必须先修改密码时同样被拦）。按模块分组，只有数量；只数令牌所属供应商自己的。
        每个模块只给有权看它的角色：没有 `area.read` 的角色拿到的 `areas` 是 null。以后每个模块加一组，不改已有的字段。
      security:
        - tenantToken: []
      responses:
        "200":
          description: 各模块的数量
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDashboardSummary"
components:
  securitySchemes:
    platformToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 平台员工登录后拿到的访问令牌
    tenantToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 租户用户登录后拿到的访问令牌；租户编号只从这里取
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      description: 每页条数；只接受十进制整数的写法
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: 上一页返回的 `next_cursor`
      schema:
        type: string
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前的版本号（记录里的 `version`），写成 `"3"` 或 `3`。没带返回 428，不是最新的返回 409
      schema:
        type: string
        examples: ['"3"']
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        幂等键：8 到 128 位的字母、数字或 `_ . : -`（用一个 UUID 就行）。同一个供应商、同一个接口、24 小时内，
        同一个键配同样的内容再来，原样返回第一次的应答，不会重复创建；配了不同的内容返回 422 `IDEMPOTENCY_KEY_REUSED`。
        失败的请求不占用键，修好以后可以带同一个键重试。没带或格式不对返回 400 `VALIDATION_FAILED`（`/idempotency-key`）。
      schema:
        type: string
        pattern: "^[A-Za-z0-9_.:-]{8,128}$"
...
  responses:
...
    ResourceNotFound:
      description: "`NOT_FOUND`：资源不存在（租户接口里，属于别的租户的资源也是这个结果）"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AreaInvalid:
      description: |
        `VALIDATION_FAILED`：字段或图形不合规。`details.issues[]` 每条有 `path`、`message`（中文），图形的问题另有机器可读的 `reason`：
        `NO_OPERATE_POLYGON`、`TOO_MANY_POLYGONS`、`TOO_MANY_TOTAL_VERTICES`（`path` 是 `/polygons`）；
        `INVALID_COORDINATE`、`TOO_FEW_POINTS`、`TOO_MANY_VERTICES`、`DUPLICATE_POINT`、`CROSSES_ANTIMERIDIAN`、`COLLINEAR`、
        `SELF_INTERSECTION`、`HAS_HOLES`（`path` 是 `/polygons/{i}/geometry`）；`RADIUS_OUT_OF_RANGE`（`/polygons/{i}/circle/radius_m`）；
        `UNKNOWN_POLYGON`（`/polygons/{i}/id`：这个编号的图形不在这个区域里）。
        部分原因带 `detail`：`SELF_INTERSECTION` 是 `{ a, b }`（第 a–a+1 个点之间的边和第 b–b+1 个点之间的边交叉，点号从 1 数）；
        `DUPLICATE_POINT` 是 `{ a, b }`；`INVALID_COORDINATE` 是 `{ point }`；`TOO_FEW_POINTS`、`TOO_MANY_VERTICES`、`TOO_MANY_POLYGONS`、`TOO_MANY_TOTAL_VERTICES` 是 `{ count }`。
        这些代码和 `@nozomi/domain` 的 `ringIssues` / `areaShapeIssues` 返回的是同一套。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    VersionRequired:
      description: "`PRECONDITION_REQUIRED`：修改时没有用请求头 `If-Match` 带版本号"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
              description: 稳定的机器可读错误码，全大写下划线
              examples: [NOT_FOUND, VALIDATION_FAILED, BAD_REQUEST, INTERNAL_ERROR]
            message:
              type: string
              description: 给人看的简体中文说明
            details:
              type: object
              description: 补充信息；没有时为空对象
              additionalProperties: true
...
    MasterDataStatus:
      type: string
      enum: [active, disabled]
      description: "active：启用；disabled：已停用（主数据不删除，只停用）"
...
    LocalizedText:
      type: object
      additionalProperties: false
      description: 多语言文本，键是语言代码（ISO 639-1）。只返回填了的语言
      properties:
        ja:
          type: string
        zh:
          type: string
        en:
          type: string
        ko:
          type: string
    LocalizedName:
      description: 多语言名称，至少有一种语言，每种最多 200 个字符
      allOf:
        - $ref: "#/components/schemas/LocalizedText"
      minProperties: 1
      examples:
        - { ja: 東京, zh: 东京, en: Tokyo }
    Point:
      type: object
      additionalProperties: false
      required: [lng, lat]
      description: WGS84 坐标，保留小数点后 6 位（多出的位数四舍五入）
      properties:
        lng:
          type: number
          minimum: -180
          maximum: 180
        lat:
          type: number
          minimum: -90
          maximum: 90
    Boundary:
      type: object
      required: [type, coordinates]
      description: |
        GeoJSON 的 Polygon 或 MultiPolygon（坐标顺序是 [经度, 纬度]）。每个环至少 4 个点且首尾相同，
        顶点总数最多 5000 个。
      properties:
        type:
          type: string
          enum: [Polygon, MultiPolygon]
        coordinates:
          type: array
...
    AreaBizType:
      type: string
      enum: [general, airport_transfer, point_to_point, charter]
      description: 业务类型：通用、接送机、点对点、包车。建商品时只能选到业务类型和商品品类相同的区域，或「通用」的区域
    AreaPolygonKind:
      type: string
      enum: [operate, forbid]
      description: "operate：营运区；forbid：禁行区（优先于营运区）"
    AreaPolygonSource:
      type: string
      enum: [drawn, pasted, circle]
      description: "这块图形是怎么来的：drawn 画的；pasted 粘贴的；circle 圆。只影响之后怎么编辑，不影响判断"
    AreaGeometry:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      description: GeoJSON 的 Polygon：只有一圈（不带洞）、首尾闭合、`[经度, 纬度]`、6 位小数、逆时针
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: array
            items:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
    AreaCircle:
      type: object
      additionalProperties: false
      required: [center, radius_m]
      description: 圆心和半径。只用于再次编辑；判断只认 `geometry` 里的多边形
      properties:
        center:
          type: object
          additionalProperties: false
          required: [lat, lng]
          properties:
            lat:
              type: number
              minimum: -90
              maximum: 90
            lng:
              type: number
              minimum: -180
              maximum: 180
        radius_m:
          type: integer
          minimum: 100
          maximum: 100000
          description: 半径（整数米）：0.1 到 100 公里
    AreaPolygon:
      type: object
      additionalProperties: false
      required: [id, kind, seq, label, source, circle, geometry]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/AreaPolygonKind"
        seq:
          type: integer
          minimum: 1
          description: 这一类里的序号（「营运 1」里的 1）。由后端分配，删除别的图形后不重排
        label:
          type: [string, "null"]
          maxLength: 40
          description: 备注名
        source:
          $ref: "#/components/schemas/AreaPolygonSource"
        circle:
          description: "`source` 是 `circle` 时有，其余为 null"
          oneOf:
            - $ref: "#/components/schemas/AreaCircle"
            - type: "null"
        geometry:
          $ref: "#/components/schemas/AreaGeometry"
    AreaPolygonInput:
      type: object
      required: [kind]
      description: |
        提交的一块图形。多边形：给 `geometry`（`circle` 不给或为 null）；圆：给 `circle`，不用给 `geometry`（给了也不看，后端按圆心半径重算）。
        已有的图形带 `id`，新加的不带。`seq` 不用给（给了被忽略）。
      properties:
        id:
          type: [string, "null"]
          format: uuid
        kind:
          $ref: "#/components/schemas/AreaPolygonKind"
        label:
          type: [string, "null"]
          maxLength: 40
        source:
          description: 不给时：有 `circle` 的是 `circle`，其余是 `drawn`
          oneOf:
            - $ref: "#/components/schemas/AreaPolygonSource"
            - type: "null"
        circle:
          oneOf:
            - $ref: "#/components/schemas/AreaCircle"
            - type: "null"
        geometry:
          description: 提交时可以是顺时针、可以多于 6 位小数、可以不闭合；带洞的不接受
          oneOf:
            - type: object
              required: [type, coordinates]
              properties:
                type:
                  type: string
                  enum: [Polygon]
                coordinates:
                  type: array
            - type: "null"
    AreaCity:
      type: object
      additionalProperties: false
      required: [id, code, name, status, center, boundary]
      description: 区域所属城市的信息（读取时现查的最新值）
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          $ref: "#/components/schemas/LocalizedName"
        status:
          $ref: "#/components/schemas/MasterDataStatus"
        center:
          $ref: "#/components/schemas/Point"
        boundary:
          oneOf:
            - $ref: "#/components/schemas/Boundary"
            - type: "null"
    AreaSummary:
      type: object
      additionalProperties: false
      description: 列表里的一项：不带图形的坐标
      required: [id, name, city_id, city, biz_type, status, operate_polygon_count, forbid_polygon_count, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          description: 多语言名称，至少一种语言，每种最多 100 个字
          allOf:
            - $ref: "#/components/schemas/LocalizedText"
        city_id:
          type: string
          format: uuid
          description: 所属城市；创建后不能改
        city:
          $ref: "#/components/schemas/AreaCity"
        biz_type:
          $ref: "#/components/schemas/AreaBizType"
        status:
          $ref: "#/components/schemas/MasterDataStatus"
        operate_polygon_count:
          type: integer
          minimum: 0
        forbid_polygon_count:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Area:
      type: object
      additionalProperties: false
      description: 区域的详情：比列表项多全部图形
      required: [id, name, city_id, city, biz_type, status, operate_polygon_count, forbid_polygon_count, version, created_at, updated_at, polygons]
      properties:
        id:
          type: string
          format: uuid
        name:
          allOf:
            - $ref: "#/components/schemas/LocalizedText"
        city_id:
          type: string
          format: uuid
        city:
          $ref: "#/components/schemas/AreaCity"
        biz_type:
          $ref: "#/components/schemas/AreaBizType"
        status:
          $ref: "#/components/schemas/MasterDataStatus"
        operate_polygon_count:
          type: integer
          minimum: 0
        forbid_polygon_count:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
          description: 版本号，每改一次加一；修改时放在请求头 `If-Match` 里
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        polygons:
          type: array
          description: 全部图形，按加入的先后
          items:
            $ref: "#/components/schemas/AreaPolygon"
    AreaCreate:
      type: object
      required: [city_id, name, biz_type, polygons]
      properties:
        city_id:
          type: string
          format: uuid
        name:
          allOf:
            - $ref: "#/components/schemas/LocalizedText"
        biz_type:
          $ref: "#/components/schemas/AreaBizType"
        polygons:
          type: array
          items:
            $ref: "#/components/schemas/AreaPolygonInput"
    AreaUpdate:
      type: object
      required: [name, biz_type, polygons]
      description: 整体替换。`city_id` 不用带；带了且和现有的不同返回 409 `FIELD_LOCKED`
      properties:
        city_id:
          type: string
          format: uuid
        name:
          allOf:
            - $ref: "#/components/schemas/LocalizedText"
        biz_type:
          $ref: "#/components/schemas/AreaBizType"
        polygons:
          type: array
          items:
            $ref: "#/components/schemas/AreaPolygonInput"
    AreaPointCheck:
      type: object
      additionalProperties: false
      required: [result, operate_polygon_ids, forbid_polygon_ids]
      properties:
        result:
          type: string
          enum: [operate, forbid, outside]
          description: "operate：在营运区里且不在禁行区里；forbid：在任何一块禁行区里；outside：都不在"
        operate_polygon_ids:
          type: array
          description: 这个位置落在其中的全部营运区（`result` 是 `forbid` 时也列出来）
          items:
            type: string
            format: uuid
        forbid_polygon_ids:
          type: array
          items:
            type: string
            format: uuid
    MapConfig:
      type: object
      additionalProperties: false
      required: [tiles]
      properties:
        tiles:
          type: [object, "null"]
          description: 栅格瓦片底图；这个环境没有配置时为 null
          additionalProperties: false
          required: [url_template, dark_url_template, min_zoom, max_zoom, tile_size, referrer_policy, attribution]
          properties:
            url_template:
              type: string
              description: 瓦片地址模板，含 `{z}`、`{x}`、`{y}`
              examples: ["https://tile.openstreetmap.org/{z}/{x}/{y}.png"]
            dark_url_template:
              type: [string, "null"]
              description: 暗色主题用的瓦片地址；没有时为 null（前端用滤镜压暗）
            min_zoom:
              type: integer
            max_zoom:
              type: integer
            tile_size:
              type: integer
              enum: [256, 512]
            referrer_policy:
              type: string
              enum: [no-referrer, origin, strict-origin, strict-origin-when-cross-origin]
              description: 请求瓦片图片时用的来源页策略（设在瓦片图片上，不改全站的策略）
            attribution:
              type: array
              description: 必须在地图上常驻显示的版权署名
              items:
                type: object
                additionalProperties: false
                required: [text, href]
                properties:
                  text:
                    type: string
                  href:
                    type: [string, "null"]
    TenantDashboardSummary:
      type: object
      additionalProperties: false
      required: [areas]
      properties:
        areas:
          type: [object, "null"]
          description: 本供应商的区域数量；当前角色没有 `area.read` 时为 null
          additionalProperties: false
          required: [active, disabled]
          properties:
            active:
              type: integer
              minimum: 0
            disabled:
              type: integer
              minimum: 0